Create ReadMe.md
# 🐱🐶 Cat vs Dog Image Classifier

A deep learning project that uses a Convolutional Neural Network (CNN) to classify images of cats and dogs. This model was trained on image data using TensorFlow and Keras, with high accuracy on validation images.

## 📌 Overview

This project aims to build an image classification model that can distinguish between images of cats and dogs. It leverages the power of CNNs and transfer learning to achieve high prediction performance.