Add files via upload
# 👤 Gender Detection from Images using Deep Learning

A deep learning project that predicts the **gender (Male/Female)** of a person from a face image using a Convolutional Neural Network (CNN). This model is trained using labeled face images and achieves high accuracy on unseen data.

---

## 📌 Overview

This project uses deep learning techniques to classify face images as either  Male  or  Female. It is built using TensorFlow and Keras and trained on a labeled dataset of human faces.